Migrated functions to functions.rs
Big refactorization
