+++ pallets/nbv-storage/src/types.rs
@@ -1,6 +1,10 @@
 use sp_core::crypto::KeyTypeId;
-
+/*--- Constants section ---*/
+pub const BDK_SERVICES_URL: &[u8] = b"http://127.0.0.1:8000";
+pub const UNSIGNED_TXS_PRIORITY: u64 = 100;
 pub const KEY_TYPE: KeyTypeId = KeyTypeId(*b"bdks");
+
+/*--- Crypto module section---*/
 pub mod crypto {
 	use super::KEY_TYPE;
 	use sp_core::sr25519::Signature as Sr25519Signature;
